op.c: Allow COW with folded constants under MAD
Under MAD builds, this copy of the SV can result in a redundant
string copy.  Since we can take advantage of COW, let’s do so.
This really speeds up compilation of things like "a"x2**31
under MAD.

XXX This does not work because of the broken way the pads ‘work’;
it has to wait until after the sprout/padconst branch is merged.